Revert previous commit for bool size.
using int makes pips maintenance easier.


git-svn-id: http://svn.cri.ensmp.fr/svn/linear/trunk@1388 8716bff9-3cf5-0310-8878-b27a67f92df8

+++ src/arithmetique/boolean.h
@@ -31,7 +31,7 @@
     #error newgen header not compatible with stdbool.h and linear header have to be compatible with newgen
 #endif
 
-typedef intptr_t bool; /* we cannot use an enum or stdbool because we need to be compatible with newgen */
+typedef int bool; /* we cannot use an enum or stdbool because we need to be compatible with newgen */
 
 #define false 0
 #define true 1
